% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_names.R
\name{b_names}
\alias{b_names}
\title{Show the column names for a data.frame in alphabetical order}
\usage{
b_names(df)
}
\arguments{
\item{df}{the data.frame to inspect}
}
\description{
This is a slight modification of \code{names}
}
\examples{
b_names(mtcars)
}
