% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm.R
\name{bm.mvtiprate}
\alias{bm.mvtiprate}
\title{Compute multivariate tip rates using phylogenetic independent contrasts}
\usage{
bm.mvtiprate(x, phy)
}
\arguments{
\item{x}{A matrix (with row names) of character state data having type \code{numeric}.}

\item{phy}{An object of class \code{tree}.}
}
\value{
A list of matrix-valued tip rates
}
\description{
As the number of terminal nodes in the phylogeny grows large the
average tip rate converges on the maximum likelihood estimate of the
Brownian motion covariance matrix
}
