% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm.R
\name{bm.tiprate}
\alias{bm.tiprate}
\title{Compute tip rates using phylogenetic independent contrasts}
\usage{
bm.tiprate(x, phy)
}
\arguments{
\item{x}{A named vector of character state data having type \code{numeric}.}

\item{phy}{An object of class \code{tree}.}
}
\value{
A vector of tip rates
}
\description{
As the number of terminal nodes in the phylogeny grows large the
average tip rate converges on the maximum likelihood estimate of the
Brownian motion variance parameter.
}
\examples{
data(squamatatree)
data(squamatamass)
phy = read.newick(text=squamatatree)
x = bm.tiprate(squamatamass, phy)
# average tiprate
mean(x) # 0.02721007
# maximum likelihood rate estimate
attr(bm.pic(squamatamass, phy)[[1]], "rate") # 0.02721699
}
