% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chelonia}
\alias{chelonia}
\title{Turtle body sizes and phylogeny}
\format{
An object of class \code{list} of length 2.
}
\source{
Eastman, J.M., Alfara, M.E., Joyce, P., Hipp, A.L., and L.J. Harmon. 2011. Evolution 65(12): 3578-3589
}
\usage{
chelonia
}
\description{
A list with turtle body sizes and a time-calibrated phylogeny of turtles.
}
\keyword{datasets}
