% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM.R
\name{ff}
\alias{ff}
\title{Creates The Use of Marginal Distributions in Conditional Forecasting}
\usage{
ff(dt)
}
\arguments{
\item{dt}{The parameter that containing the data}
}
\description{
Creates The Use of Marginal Distributions in Conditional Forecasting
}
