% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squamatamass}
\alias{squamatamass}
\title{Squamate reptile body mass}
\format{
A named vector of class \code{numeric} with body mass data for
for 3937 squamates.
}
\source{
Feldman, A., N. Sabath, R.A. Pyron, I. Mayrose, and S. Meiri. 2016.
Global Ecology and Biogeography 25(2): 187-197
}
\usage{
squamatamass
}
\description{
A dataset from Feldman et al. (2016) containing log10-transformed body masses
(in grams) of lizards and snakes.
}
\keyword{datasets}
