% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{squamatatree}
\alias{squamatatree}
\title{Squamate reptile phylogeny}
\format{
A Newick formatted character string
}
\source{
Tonini, J., Beard, K.H., Ferreira, R.B., Jetz, W., and R.A. Pyron. 2016. Biological Conservation 204: 23-31
}
\usage{
squamatatree
}
\description{
A time-calibrated phylogeny of squamate reptiles.
}
\keyword{datasets}
